+++ CHANGES.md
@@ -21,12 +21,15 @@ of that change.
 - `batch_size` property in DeviceDataLoader
 - `ClassificationLearner.predict` to get prediction on a single item
 - Monkey-patched torch.Tensor so matplotlib works
+- Data block API
+- Monkey-patched torch.utils.data.dataloader.DataLoader to create a passthrough to the dataset
 
 ### Changed:
 
 - `validate` now takes optional `n_batch`
 - `create_cnn` now returns a `ClassificationLearner`
 - `return_path` flag to `Learner.save`
+- `ImageDataBunch.show_batch()` now works for every type of dataset, removes `show_images` and `show_xy_images` as a result.
 
 ### Fixed: